prometheusremotewrite: release lock when wal empty
Previously, the wal tailing routine held an exclusive lock over the
entire wal, while waiting for new sections to be written.

This led to a deadlock situtation, as a locked wal naturally can not be written to.

To remedy this, the lock is now only held during actual read attempts,
not in between.

Furthermore, inotify watching of the wal dir has been replaced with
channel-based signaling between the producers and the tailing routine,
so that the latter blocks until any new writes have happened.

prometheusremotewrite: wal default nop-logger

Initializes the WAL with a default nop logger, instead of previous nil
which led to panics.

doc: add changelog entry

make gotidy

Signed-off-by: Benedikt Bongartz <bongartz@klimlive.de>